Do not print redundant 'pass' for classes, hide __module__
Ensure that classes are always separated by an extra new line, not only
when 'pass' is present. Hide 'pass' when a docstring is available for
the class.

Spit out less useless newlines.

Hide __module__ which seems to be added automatically for every class.

Added tests to verify the expected output for empty classes and
classes with a single docstring.