Fix missing keyword argument for `__new__` methods
`inspect.signature` seems to have even more edge cases, like dropping
the first parameter for the `__new__` method. Instead of handling every
case individually, let's rely on heuristics: if the first argument name
from the known correct `getfullargspec` output is missing in the
signature, assume that it was dropped and restore it.